调整buffer的refcnt字段类型避免填充 event_init传入的最大事件数只限制io事件，每次event_wait处理所有超时的timer stream每次读写完都检查error标志，不用等到下次调用

+++ include/cbase/buffer.h
@@ -17,7 +17,7 @@ struct buffer {
 	 */
 	char *head, *data, *tail, *end;	/* position pointers */
 	struct buffer *parent;	/* if separated from another buffer */
-	unsigned int refcnt;	/* ref counter */
+	unsigned long refcnt;	/* ref counter */
 	struct list_head list;	/* join with other buffers into a chain */
 	struct pool *pool;
 };
